+++ src/main/java/chess/domain/board/Initiator.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.board;
 
+import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 
 import java.util.Map;
