+++ src/main/java/chess/domain/piece/role/Knight.java
@@ -1,4 +1,4 @@
-package chess.domain.pieces;
+package chess.domain.piece.role;
 
 import chess.domain.position.Position;
