+++ src/test/java/chess/domain/BoardTest.java
@@ -1,6 +1,14 @@
 package chess.domain;
 
-import chess.domain.pieces.*;
+import chess.domain.board.Board;
+import chess.domain.board.BoardInitiator;
+import chess.domain.piece.*;
+import chess.domain.piece.role.Bishop;
+import chess.domain.piece.role.King;
+import chess.domain.piece.role.Knight;
+import chess.domain.piece.role.Pawn;
+import chess.domain.piece.role.Queen;
+import chess.domain.piece.role.Rook;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
